requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
bdd==0.2
behave==1.2.5
decorator==4.1.1
diff-match-patch==20121119
Django==1.11
django-adminplus==0.5
django-behave==0.1.5
django-bootstrap-pagination==1.6.2
django-debug-toolbar==1.8
django-extensions==1.8.1
django-filter==1.0.4
django-import-export==0.5.1
django-model-utils==3.0.0
django-mysql==2.1.0
django-querycount==0.6.0
enum34==1.1.6
et-xmlfile==1.0.1
factory-boy==2.8.1
Faker==0.7.18
ipaddress==1.0.18
ipython==5.4.1
ipython-genutils==0.2.0
jdcal==1.3
mysqlclient==1.3.10
odfpy==1.3.5
openpyxl==2.4.8
parse==1.8.2
parse-type==0.3.4
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
scandir==1.5
selenium==3.4.3
simplegeneric==0.8.1
six==1.10.0
splinter==0.7.5
sqlparse==0.2.3
tablib==0.11.5
traitlets==4.3.2
unicodecsv==0.14.1
wcwidth==0.1.7
xlrd==1.0.0
xlwt==1.2.0
